refactor: Update antd dependency to version "^5.20.3"
The code changes in 'package.json' update the antd dependency to version "^5.20.3". This update allows the application to utilize the latest features and bug fixes provided by the antd library for UI components. Updating the dependency enhances the user interface with pre-built, customizable components and ensures compatibility with the latest version of antd.